moved get_distance_function from utils to distance.h/cpp
