+++ src/distance.cpp
@@ -15,12 +15,16 @@
 #include <iostream>
 
 #include "distance.h"
+#include "utils.h"
 #include "logger.h"
 #include "ann_exception.h"
 
 namespace diskann {
 
-  // Cosine similarity.
+  //
+  // Cosine distance functions.
+  //
+
   float DistanceCosineInt8::compare(const int8_t *a, const int8_t *b,
                                     uint32_t length) const {
 #ifdef _WINDOWS
@@ -65,7 +69,10 @@ namespace diskann {
     return 1.0f - (float) (scalarProduct / (sqrt(magA) * sqrt(magB)));
   }
 
+  //
   // L2 distance functions.
+  //
+  
   float DistanceL2Int8::compare(const int8_t *a, const int8_t *b,
                                 uint32_t size) const {
     int32_t result = 0;
@@ -498,6 +505,109 @@ namespace diskann {
              unpack[5] + unpack[6] + unpack[7];
 
     return -result;
+  }
+
+  // Get the right distance function for the given metric.
+  template<>
+  diskann::Distance<float> *get_distance_function(diskann::Metric m) {
+    if (m == diskann::Metric::L2) {
+      if (Avx2SupportedCPU) {
+        diskann::cout << "L2: Using AVX2 distance computation DistanceL2Float"
+                      << std::endl;
+        return new diskann::DistanceL2Float();
+      } else if (AvxSupportedCPU) {
+        diskann::cout
+            << "L2: AVX2 not supported. Using AVX distance computation"
+            << std::endl;
+        return new diskann::AVXDistanceL2Float();
+      } else {
+        diskann::cout << "L2: Older CPU. Using slow distance computation"
+                      << std::endl;
+        return new diskann::SlowDistanceL2Float();
+      }
+    } else if (m == diskann::Metric::COSINE) {
+      diskann::cout << "Cosine: Using either AVX or AVX2 implementation"
+                    << std::endl;
+      return new diskann::DistanceCosineFloat();
+    } else if (m == diskann::Metric::INNER_PRODUCT) {
+      diskann::cout << "Inner product: Using AVX2 implementation "
+                       "AVXDistanceInnerProductFloat"
+                    << std::endl;
+      return new diskann::AVXDistanceInnerProductFloat();
+    } else if (m == diskann::Metric::FAST_L2) {
+      diskann::cout << "Fast_L2: Using AVX2 implementation with norm "
+                       "memoization DistanceFastL2<float>"
+                    << std::endl;
+      return new diskann::DistanceFastL2<float>();
+    } else {
+      std::stringstream stream;
+      stream << "Only L2, cosine, and inner product supported for floating "
+                "point vectors as of now."
+             << std::endl;
+      diskann::cerr << stream.str() << std::endl;
+      throw diskann::ANNException(stream.str(), -1, __FUNCSIG__, __FILE__,
+                                  __LINE__);
+    }
+  }
+
+  template<>
+  diskann::Distance<int8_t> *get_distance_function(diskann::Metric m) {
+    if (m == diskann::Metric::L2) {
+      if (Avx2SupportedCPU) {
+        diskann::cout << "Using AVX2 distance computation DistanceL2Int8."
+                      << std::endl;
+        return new diskann::DistanceL2Int8();
+      } else if (AvxSupportedCPU) {
+        diskann::cout << "AVX2 not supported. Using AVX distance computation"
+                      << std::endl;
+        return new diskann::AVXDistanceL2Int8();
+      } else {
+        diskann::cout << "Older CPU. Using slow distance computation "
+                         "SlowDistanceL2Int<int8_t>."
+                      << std::endl;
+        return new diskann::SlowDistanceL2Int<int8_t>();
+      }
+    } else if (m == diskann::Metric::COSINE) {
+      diskann::cout << "Using either AVX or AVX2 for Cosine similarity "
+                       "DistanceCosineInt8."
+                    << std::endl;
+      return new diskann::DistanceCosineInt8();
+    } else {
+      std::stringstream stream;
+      stream << "Only L2 and cosine supported for signed byte vectors."
+             << std::endl;
+      diskann::cerr << stream.str() << std::endl;
+      throw diskann::ANNException(stream.str(), -1, __FUNCSIG__, __FILE__,
+                                  __LINE__);
+    }
+  }
+
+  template<>
+  diskann::Distance<uint8_t> *get_distance_function(diskann::Metric m) {
+    if (m == diskann::Metric::L2) {
+#ifdef _WINDOWS
+      diskann::cout
+          << "WARNING: AVX/AVX2 distance function not defined for Uint8. Using "
+             "slow version. "
+             "Contact gopalsr@microsoft.com if you need AVX/AVX2 support."
+          << std::endl;
+#endif
+      return new diskann::DistanceL2UInt8();
+    } else if (m == diskann::Metric::COSINE) {
+      diskann::cout
+          << "AVX/AVX2 distance function not defined for Uint8. Using "
+             "slow version SlowDistanceCosineUint8() "
+             "Contact gopalsr@microsoft.com if you need AVX/AVX2 support."
+          << std::endl;
+      return new diskann::SlowDistanceCosineUInt8();
+    } else {
+      std::stringstream stream;
+      stream << "Only L2 and cosine supported for unsigned byte vectors."
+             << std::endl;
+      diskann::cerr << stream.str() << std::endl;
+      throw diskann::ANNException(stream.str(), -1, __FUNCSIG__, __FILE__,
+                                  __LINE__);
+    }
   }
 
   template DISKANN_DLLEXPORT class DistanceInnerProduct<float>;
